Moved db models to new package
