+++ database/model/stock.go
